+++ docs/CHANGELOG.rst
@@ -20,6 +20,7 @@ Added
  - Docker executor
  - Celery engine
  - Purge command
+ - Framework for testing processors
 
 
 ==================